+++ tripartite-ding/tripartite-ding-openapi/src/main/java/work/gaigeshen/tripartite/ding/openapi/parameters/process/DingProcessWorkRecordBatchUpdateParameters.java
@@ -21,7 +21,7 @@ public class DingProcessWorkRecordBatchUpdateParameters implements ClientParamet
 
         public Collection<Instance> instances;
 
-        public Integer agentid;
+        public Long agentid;
     }
 
     public static class Instance {
