+++ tripartite-ding/tripartite-ding-openapi/src/main/java/work/gaigeshen/tripartite/ding/openapi/parameters/process/DingProcessWorkRecordTaskCancelParameters.java
@@ -19,7 +19,7 @@ public class DingProcessWorkRecordTaskCancelParameters implements ClientParamete
 
     public static class Request {
 
-        public Integer agentid;
+        public Long agentid;
 
         public String process_instance_id;
